% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_addition.R
\name{cppsum}
\alias{cppsum}
\title{Add two vectors using cpp integrated with R.}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add two vectors using cpp integrated with R.
}
\examples{
cppsum(c(1,3,4), c(3,2,6))
}

