% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demovector.R
\docType{package}
\name{demovector}
\alias{demovector}
\alias{demovector-package}
\title{demovector: A package for adding two vectors using Rcpp integration}
\description{
The package provides a function for addition of vectors: cppsum
}

